Move tests into the module under test.
