+++ src/structs/reqs.rs
@@ -1,5 +1,7 @@
-use uuid::Uuid;
 use std::borrow::Cow;
+use uuid::Uuid;
+
+use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct Order<'a> {
@@ -45,11 +47,7 @@ pub enum MarketType {
 }
 
 impl<'a> Order<'a> {
-    pub fn market<T: Into<Cow<'a, str>>>(
-        product_id: T,
-        side: OrderSide,
-        size: f64,
-    ) -> Self {
+    pub fn market<T: Into<Cow<'a, str>>>(product_id: T, side: OrderSide, size: f64) -> Self {
         Order {
             product_id: product_id.into(),
             client_oid: None,
@@ -74,7 +72,7 @@ impl<'a> Order<'a> {
         side: OrderSide,
         size: f64,
         price: f64,
-        post_only: bool
+        post_only: bool,
     ) -> Self {
         Order {
             product_id: product_id.into(),
@@ -90,22 +88,35 @@ impl<'a> Order<'a> {
         }
     }
 
-    pub fn buy_limit<T: Into<Cow<'a, str>>>(product_id: T, size: f64, price: f64, post_only: bool) -> Self {
+    pub fn buy_limit<T: Into<Cow<'a, str>>>(
+        product_id: T,
+        size: f64,
+        price: f64,
+        post_only: bool,
+    ) -> Self {
         Self::limit(product_id, OrderSide::Buy, size, price, post_only)
     }
 
-    pub fn sell_limit<T: Into<Cow<'a, str>>>(product_id: T, size: f64, price: f64, post_only: bool) -> Self {
+    pub fn sell_limit<T: Into<Cow<'a, str>>>(
+        product_id: T,
+        size: f64,
+        price: f64,
+        post_only: bool,
+    ) -> Self {
         Self::limit(product_id, OrderSide::Sell, size, price, post_only)
     }
 
     pub fn client_oid(self, client_oid: Uuid) -> Self {
         let client_oid = Some(client_oid);
-        Order{client_oid, .. self }
+        Order { client_oid, ..self }
     }
 
     pub fn stop(self, price: f64, stop_type: OrderStopType) -> Self {
-        let stop = Some(OrderStop { stop_price: price, _type: stop_type });
-        Order{stop, .. self}
+        let stop = Some(OrderStop {
+            stop_price: price,
+            _type: stop_type,
+        });
+        Order { stop, ..self }
     }
 
     pub fn stop_loss(self, price: f64) -> Self {
@@ -118,14 +129,23 @@ impl<'a> Order<'a> {
 
     pub fn time_in_force(self, time_in_force: OrderTimeInForce) -> Self {
         match self._type {
-            OrderType::Limit {price, size, post_only, ..} => {
-                let _type = OrderType::Limit {price, size, post_only, time_in_force: Some(time_in_force)};
-                Order{_type, .. self}
-            },
-            _ => panic!("time_in_force is for limit orders only")
+            OrderType::Limit {
+                price,
+                size,
+                post_only,
+                ..
+            } => {
+                let _type = OrderType::Limit {
+                    price,
+                    size,
+                    post_only,
+                    time_in_force: Some(time_in_force),
+                };
+                Order { _type, ..self }
+            }
+            _ => panic!("time_in_force is for limit orders only"),
         }
     }
-
 }
 
 #[derive(Serialize, Deserialize, Debug)]
@@ -171,8 +191,11 @@ mod tests {
         assert!(o.client_oid.is_none());
 
         match &o._type {
-            OrderType::Limit {time_in_force: None, ..} => assert!(true),
-            _ => assert!(false)
+            OrderType::Limit {
+                time_in_force: None,
+                ..
+            } => assert!(true),
+            _ => assert!(false),
         }
 
         let o = Order::buy_limit("BTC-USD", 10.0, 100.0, true)
@@ -183,15 +206,24 @@ mod tests {
         assert!(o.stop.is_some());
 
         match &o._type {
-            OrderType::Limit {time_in_force: Some(OrderTimeInForce::GTC), ..} => assert!(true),
-            _ => assert!(false)
+            OrderType::Limit {
+                time_in_force: Some(OrderTimeInForce::GTC),
+                ..
+            } => assert!(true),
+            _ => assert!(false),
         }
     }
 
     #[derive(Debug)]
-    enum Coin { AAA, BBB}
+    enum Coin {
+        AAA,
+        BBB,
+    }
     #[derive(Debug)]
-    struct Pair {a: Coin, b:Coin}
+    struct Pair {
+        a: Coin,
+        b: Coin,
+    }
 
     impl<'a> From<Pair> for Order<'a> {
         fn from(pair: Pair) -> Self {
@@ -201,7 +233,10 @@ mod tests {
 
     #[test]
     fn test_order_from() {
-        let p = Pair {a: Coin::AAA, b:Coin::BBB};
+        let p = Pair {
+            a: Coin::AAA,
+            b: Coin::BBB,
+        };
         let order_owned: Order = p.into();
         assert_eq!(order_owned.product_id, "AAA-BBB");
         let order_str: Order = Order::buy_market("AAA-BBB", 10.0);
